Crate google-spectrum1_explorer

This documentation was generated from spectrum crate version 0.0.1. The original source code can be found on github.

Features

It seems there is nothing you can do here ... .

paws (get spectrum, get spectrum batch, init, notify spectrum use, register and verify device)

Everything else about the spectrum v1_explorer API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.paws().get_spectrum_batch(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-spectrum1_explorer = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-spectrum1_explorer" as spectrum1_explorer;
use spectrum1_explorer::PawsGetSpectrumBatchRequest;
use spectrum1_explorer::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use spectrum1_explorer::Spectrum;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Spectrum::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: PawsGetSpectrumBatchRequest = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.paws().get_spectrum_batch(&req)
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Structs

AntennaCharacteristics
Antenna characteristics provide additional information, such as the antenna height, antenna type, etc. Whether antenna characteristics must be provided in a request depends on the device type and regulatory domain.

This type is not used in any activity, and only used as part of another schema.

DatabaseSpec
This message contains the name and URI of a database.

This type is not used in any activity, and only used as part of another schema.

DbUpdateSpec
This message is provided by the database to notify devices of an upcoming change to the database URI.

This type is not used in any activity, and only used as part of another schema.

DeviceCapabilities
Device capabilities provide additional information that may be used by a device to provide additional information to the database that may help it to determine available spectrum. If the database does not support device capabilities it will ignore the parameter altogether.

This type is not used in any activity, and only used as part of another schema.

DeviceDescriptor
The device descriptor contains parameters that identify the specific device, such as its manufacturer serial number, regulatory-specific identifier (e.g., FCC ID), and any other device characteristics required by regulatory domains.

This type is not used in any activity, and only used as part of another schema.

DeviceOwner
This parameter contains device-owner information required as part of device registration. The regulatory domains may require additional parameters.

All contact information must be expressed using the structure defined by the vCard format specification. Only the contact fields of vCard are supported:
- fn: Full name of an individual - org: Name of the organization - adr: Address fields - tel: Telephone numbers - email: Email addresses

Note that the vCard specification defines maximum lengths for each field.

This type is not used in any activity, and only used as part of another schema.

DeviceValidity
The device validity element describes whether a particular device is valid to operate in the regulatory domain.

This type is not used in any activity, and only used as part of another schema.

EventTime
The start and stop times of an event. This is used to indicate the time period for which a spectrum profile is valid.

Both times are expressed using the format, YYYY-MM-DDThh:mm:ssZ, as defined in RFC3339. The times must be expressed using UTC.

This type is not used in any activity, and only used as part of another schema.

FrequencyRange
A specific range of frequencies together with the associated maximum power level and channel identifier.

This type is not used in any activity, and only used as part of another schema.

GeoLocation
This parameter is used to specify the geolocation of the device.

This type is not used in any activity, and only used as part of another schema.

GeoLocationEllipse
A "point" with uncertainty is represented using the Ellipse shape.

This type is not used in any activity, and only used as part of another schema.

GeoLocationPoint
A single geolocation on the globe.

This type is not used in any activity, and only used as part of another schema.

GeoLocationPolygon
A region is represented using the polygonal shape.

This type is not used in any activity, and only used as part of another schema.

GeoSpectrumSchedule
The schedule of spectrum profiles available at a particular geolocation.

This type is not used in any activity, and only used as part of another schema.

PawGetSpectrumBatchMethodBuilder
The Google Spectrum Database does not support batch requests, so this method always yields an UNIMPLEMENTED error.

PawGetSpectrumMethodBuilder
Requests information about the available spectrum for a device at a location. Requests from a fixed-mode device must include owner information so the device can be registered with the database.

PawInitMethodBuilder
Initializes the connection between a white space device and the database.

PawMethodsBuilder
A builder providing access to all methods supported on paw resources. It is not used directly, but through the Spectrum hub.

PawNotifySpectrumUseMethodBuilder
Notifies the database that the device has selected certain frequency ranges for transmission. Only to be invoked when required by the regulator. The Google Spectrum Database does not operate in domains that require notification, so this always yields an UNIMPLEMENTED error.

PawRegisterMethodBuilder
The Google Spectrum Database implements registration in the getSpectrum method. As such this always returns an UNIMPLEMENTED error.

PawVerifyDeviceMethodBuilder
Validates a device for white space use in accordance with regulatory rules. The Google Spectrum Database does not support master/slave configurations, so this always yields an UNIMPLEMENTED error.

PawsGetSpectrumBatchRequest
The request message for a batch available spectrum query protocol.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get spectrum batch paws (request)

PawsGetSpectrumBatchResponse
The response message for the batch available spectrum query contains a schedule of available spectrum for the device at multiple locations.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get spectrum batch paws (response)

PawsGetSpectrumRequest
The request message for the available spectrum query protocol which must include the device's geolocation.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get spectrum paws (request)

PawsGetSpectrumResponse
The response message for the available spectrum query which contains a schedule of available spectrum for the device.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get spectrum paws (response)

PawsInitRequest
The initialization request message allows the master device to initiate exchange of capabilities with the database.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

init paws (request)

PawsInitResponse
The initialization response message communicates database parameters to the requesting device.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

init paws (response)

PawsNotifySpectrumUseRequest
The spectrum-use notification message which must contain the geolocation of the Device and parameters required by the regulatory domain.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

notify spectrum use paws (request)

PawsNotifySpectrumUseResponse
An empty response to the notification.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

notify spectrum use paws (response)

PawsRegisterRequest
The registration request message contains the required registration parameters.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

register paws (request)

PawsRegisterResponse
The registration response message simply acknowledges receipt of the request and is otherwise empty.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

register paws (response)

PawsVerifyDeviceRequest
The device validation request message.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

verify device paws (request)

PawsVerifyDeviceResponse
The device validation response message.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

verify device paws (response)

RulesetInfo
This contains parameters for the ruleset of a regulatory domain that is communicated using the initialization and available-spectrum processes.

This type is not used in any activity, and only used as part of another schema.

Spectrum
Central instance to access all Spectrum related resource activities

SpectrumMessage
Available spectrum can be logically characterized by a list of frequency ranges and permissible power levels for each range.

This type is not used in any activity, and only used as part of another schema.

SpectrumSchedule
The spectrum schedule element combines an event time with spectrum profile to define a time period in which the profile is valid.

This type is not used in any activity, and only used as part of another schema.

Vcard
A vCard-in-JSON message that contains only the fields needed for PAWS: - fn: Full name of an individual - org: Name of the organization - adr: Address fields - tel: Telephone numbers - email: Email addresses

This type is not used in any activity, and only used as part of another schema.

VcardAddress
The structure used to represent a street address.

This type is not used in any activity, and only used as part of another schema.

VcardTelephone
The structure used to represent a telephone number.

This type is not used in any activity, and only used as part of another schema.

VcardTypedText
The structure used to represent an organization and an email address.

This type is not used in any activity, and only used as part of another schema.