Struct google_spectrum1_explorer::DbUpdateSpec

pub struct DbUpdateSpec {
    pub databases: Option<Vec<DatabaseSpec>>,
}

This message is provided by the database to notify devices of an upcoming change to the database URI.

This type is not used in any activity, and only used as part of another schema.

Fields
databases: Option<Vec<DatabaseSpec>>

A required list of one or more databases. A device should update its preconfigured list of databases to replace (only) the database that provided the response with the specified entries.