Struct google_spectrum1_explorer::DeviceDescriptor

pub struct DeviceDescriptor {
    pub etsi_en_device_type: Option<String>,
    pub fcc_id: Option<String>,
    pub serial_number: Option<String>,
    pub etsi_en_technology_id: Option<String>,
    pub fcc_tvbd_device_type: Option<String>,
    pub etsi_en_device_category: Option<String>,
    pub ruleset_ids: Option<Vec<String>>,
    pub etsi_en_device_emissions_class: Option<String>,
    pub manufacturer_id: Option<String>,
    pub model_id: Option<String>,
}

The device descriptor contains parameters that identify the specific device, such as its manufacturer serial number, regulatory-specific identifier (e.g., FCC ID), and any other device characteristics required by regulatory domains.

This type is not used in any activity, and only used as part of another schema.

Fields

etsi_en_device_type

Specifies the ETSI white space device type. Valid values are single-letter strings, such as A, B, etc. Consult the ETSI documentation for details about the device types.

fcc_id

Specifies the device's FCC certification identifier. The value is an identifier string whose length should not exceed 32 characters. Note that, in practice, a valid FCC ID may be limited to 19 characters.

serial_number

The manufacturer's device serial number; required by the applicable regulatory domain. The length of the value must not exceed 64 characters.

etsi_en_technology_id

Specifies the ETSI white space device technology identifier. The string value must not exceed 64 characters in length. Consult the ETSI documentation for details about the device types.

fcc_tvbd_device_type

Specifies the TV Band White Space device type, as defined by the FCC. Valid values are FIXED, MODE_1, MODE_2.

etsi_en_device_category

Specifies the ETSI white space device category. Valid values are the strings master and slave. This field is case-insensitive. Consult the ETSI documentation for details about the device types.

ruleset_ids

The list of identifiers for rulesets supported by the device. A database may require that the device provide this list before servicing the device requests. If the database does not support any of the rulesets specified in the list, the database may refuse to service the device requests. If present, the list must contain at least one entry.

For information about the valid requests, see section 9.2 of the PAWS specification. Currently, FccTvBandWhiteSpace-2010 is the only supported ruleset.

etsi_en_device_emissions_class

Specifies the ETSI white space device emissions class. The values are represented by numeric strings, such as 1, 2, etc. Consult the ETSI documentation for details about the device types.

manufacturer_id

The manufacturer's ID may be required by the regulatory domain. This should represent the name of the device manufacturer, should be consistent across all devices from the same manufacturer, and should be distinct from that of other manufacturers. The string value must not exceed 64 characters in length.

model_id

The device's model ID may be required by the regulatory domain. The string value must not exceed 64 characters in length.