Struct google_spectrum1_explorer::GeoLocationEllipse

pub struct GeoLocationEllipse {
    pub center: Option<GeoLocationPoint>,
    pub semi_minor_axis: Option<f64>,
    pub orientation: Option<f64>,
    pub semi_major_axis: Option<f64>,
}

A "point" with uncertainty is represented using the Ellipse shape.

This type is not used in any activity, and only used as part of another schema.

Fields

center

A required geo-spatial point representing the center of the ellipse.

semi_minor_axis

A floating-point number that expresses the location uncertainty along the minor axis of the ellipse. May be required by the regulatory domain. When the uncertainty is optional, the default value is 0.

orientation

A floating-point number that expresses the orientation of the ellipse, representing the rotation, in degrees, of the semi-major axis from North towards the East. For example, when the uncertainty is greatest along the North-South direction, orientation is 0 degrees; conversely, if the uncertainty is greatest along the East-West direction, orientation is 90 degrees. When orientation is not present, the orientation is assumed to be 0.

semi_major_axis

A floating-point number that expresses the location uncertainty along the major axis of the ellipse. May be required by the regulatory domain. When the uncertainty is optional, the default value is 0.