Struct google_spectrum1_explorer::PawMethods

pub struct PawMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on paw resources. It is not used directly, but through the Spectrum hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_spectrum1_explorer as spectrum1_explorer;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use spectrum1_explorer::Spectrum;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Spectrum::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get_spectrum(...)`, `get_spectrum_batch(...)`, `init(...)`, `notify_spectrum_use(...)`, `register(...)` and `verify_device(...)`
// to build up your call.
let rb = hub.paws();

Methods

impl<'a, C, A> PawMethods<'a, C, A>

fn notify_spectrum_use(&self, request: PawsNotifySpectrumUseRequest) -> PawNotifySpectrumUseCall<'a, C, A>

Create a builder to help you perform the following task:

Notifies the database that the device has selected certain frequency ranges for transmission. Only to be invoked when required by the regulator. The Google Spectrum Database does not operate in domains that require notification, so this always yields an UNIMPLEMENTED error.

Arguments

request - No description provided.

fn register(&self, request: PawsRegisterRequest) -> PawRegisterCall<'a, C, A>

Create a builder to help you perform the following task:

The Google Spectrum Database implements registration in the getSpectrum method. As such this always returns an UNIMPLEMENTED error.

Arguments

request - No description provided.

fn get_spectrum(&self, request: PawsGetSpectrumRequest) -> PawGetSpectrumCall<'a, C, A>

Create a builder to help you perform the following task:

Requests information about the available spectrum for a device at a location. Requests from a fixed-mode device must include owner information so the device can be registered with the database.

Arguments

request - No description provided.

fn init(&self, request: PawsInitRequest) -> PawInitCall<'a, C, A>

Create a builder to help you perform the following task:

Initializes the connection between a white space device and the database.

Arguments

request - No description provided.

fn get_spectrum_batch(&self, request: PawsGetSpectrumBatchRequest) -> PawGetSpectrumBatchCall<'a, C, A>

Create a builder to help you perform the following task:

The Google Spectrum Database does not support batch requests, so this method always yields an UNIMPLEMENTED error.

Arguments

request - No description provided.

fn verify_device(&self, request: PawsVerifyDeviceRequest) -> PawVerifyDeviceCall<'a, C, A>

Create a builder to help you perform the following task:

Validates a device for white space use in accordance with regulatory rules. The Google Spectrum Database does not support master/slave configurations, so this always yields an UNIMPLEMENTED error.

Arguments

request - No description provided.