Struct google_spectrum1_explorer::PawsGetSpectrumResponse

pub struct PawsGetSpectrumResponse {
    pub kind: Option<String>,
    pub device_desc: Option<DeviceDescriptor>,
    pub database_change: Option<DbUpdateSpec>,
    pub max_total_bw_hz: Option<f64>,
    pub spectrum_schedules: Option<Vec<SpectrumSchedule>>,
    pub max_contiguous_bw_hz: Option<f64>,
    pub version: Option<String>,
    pub timestamp: Option<String>,
    pub ruleset_info: Option<RulesetInfo>,
    pub needs_spectrum_report: Option<bool>,
    pub type_: Option<String>,
}

The response message for the available spectrum query which contains a schedule of available spectrum for the device.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get spectrum paws (response)

Fields

kind

Identifies what kind of resource this is. Value: the fixed string "spectrum#pawsGetSpectrumResponse".

device_desc

The database must return, in its available spectrum response, the device descriptor information it received in the master device's available spectrum request.

database_change

A database may include the databaseChange parameter to notify a device of a change to its database URI, providing one or more alternate database URIs. The device should use this information to update its list of pre-configured databases by (only) replacing its entry for the responding database with the list of alternate URIs.

max_total_bw_hz

The database may return a constraint on the allowed maximum total bandwidth (in Hertz), which need not be contiguous. A regulatory domain may require the database to return this parameter. When this parameter is present in the available spectrum response, the device must apply this constraint to its spectrum-selection logic to ensure that total bandwidth does not exceed this value.

spectrum_schedules

The available spectrum response must contain a spectrum schedule list. The list may be empty if spectrum is not available. The database may return more than one spectrum schedule to represent future changes to the available spectrum. How far in advance a schedule may be provided depends on the applicable regulatory domain.

max_contiguous_bw_hz

The database may return a constraint on the allowed maximum contiguous bandwidth (in Hertz). A regulatory domain may require the database to return this parameter. When this parameter is present in the response, the device must apply this constraint to its spectrum-selection logic to ensure that no single block of spectrum has bandwidth that exceeds this value.

version

The PAWS version. Must be exactly 1.0.

Required field.

timestamp

The database includes a timestamp of the form YYYY-MM-DDThh:mm:ssZ (Internet timestamp format per RFC3339) in its available spectrum response. The timestamp should be used by the device as a reference for the start and stop times specified in the response spectrum schedules.

ruleset_info

The database should return ruleset information, which identifies the applicable regulatory authority and ruleset for the available spectrum response. If included, the device must use the corresponding ruleset to interpret the response. Values provided in the returned ruleset information, such as maxLocationChange, take precedence over any conflicting values provided in the ruleset information returned in a prior initialization response sent by the database to the device.

needs_spectrum_report

For regulatory domains that require a spectrum-usage report from devices, the database must return true for this parameter if the spectrum schedule list is not empty; otherwise, the database will either return false or omit this parameter. If this parameter is present and its value is true, the device must send a spectrum use notify message to the database; otherwise, the device must not send the notification.

type_

The message type (e.g., INIT_REQ, AVAIL_SPECTRUM_REQ, ...).

Required field.