Struct google_spectrum1_explorer::SpectrumMessage

pub struct SpectrumMessage {
    pub bandwidth: Option<f64>,
    pub frequency_ranges: Option<Vec<FrequencyRange>>,
}

Available spectrum can be logically characterized by a list of frequency ranges and permissible power levels for each range.

This type is not used in any activity, and only used as part of another schema.

Fields

bandwidth

The bandwidth (in Hertz) for which permissible power levels are specified. For example, FCC regulation would require only one spectrum specification at 6MHz bandwidth, but Ofcom regulation would require two specifications, at 0.1MHz and 8MHz. This parameter may be empty if there is no available spectrum. It will be present otherwise.

frequency_ranges

The list of frequency ranges and permissible power levels. The list may be empty if there is no available spectrum, otherwise it will be present.