Struct google_spectrum1_explorer::SpectrumSchedule

pub struct SpectrumSchedule {
    pub event_time: Option<EventTime>,
    pub spectra: Option<Vec<SpectrumMessage>>,
}

The spectrum schedule element combines an event time with spectrum profile to define a time period in which the profile is valid.

This type is not used in any activity, and only used as part of another schema.

Fields

event_time

The event time expresses when the spectrum profile is valid. It will always be present.

spectra

A list of spectrum messages representing the usable profile. It will always be present, but may be empty when there is no available spectrum.