Crate google_spectrum1_explorer

This documentation was generated from spectrum crate version 1.0.0+20160606, where 20160606 is the exact revision of the spectrum:v1explorer schema built by the mako code generator v1.0.0.

Everything else about the spectrum v1_explorer API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

paws

get spectrum, get spectrum batch, init, notify spectrum use, register and verify device

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.paws().get_spectrum_batch(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-spectrum1_explorer = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_spectrum1_explorer as spectrum1_explorer;
use spectrum1_explorer::PawsGetSpectrumBatchRequest;
use spectrum1_explorer::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use spectrum1_explorer::Spectrum;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Spectrum::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = PawsGetSpectrumBatchRequest::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.paws().get_spectrum_batch(req)
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AntennaCharacteristics
Antenna characteristics provide additional information, such as the antenna height, antenna type, etc. Whether antenna characteristics must be provided in a request depends on the device type and regulatory domain.

DatabaseSpec
This message contains the name and URI of a database.

DbUpdateSpec
This message is provided by the database to notify devices of an upcoming change to the database URI.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DeviceCapabilities
Device capabilities provide additional information that may be used by a device to provide additional information to the database that may help it to determine available spectrum. If the database does not support device capabilities it will ignore the parameter altogether.

DeviceDescriptor
The device descriptor contains parameters that identify the specific device, such as its manufacturer serial number, regulatory-specific identifier (e.g., FCC ID), and any other device characteristics required by regulatory domains.

DeviceOwner
This parameter contains device-owner information required as part of device registration. The regulatory domains may require additional parameters.

DeviceValidity
The device validity element describes whether a particular device is valid to operate in the regulatory domain.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

EventTime
The start and stop times of an event. This is used to indicate the time period for which a spectrum profile is valid.

FrequencyRange
A specific range of frequencies together with the associated maximum power level and channel identifier.

GeoLocation
This parameter is used to specify the geolocation of the device.

GeoLocationEllipse
A "point" with uncertainty is represented using the Ellipse shape.

GeoLocationPoint
A single geolocation on the globe.

GeoLocationPolygon
A region is represented using the polygonal shape.

GeoSpectrumSchedule
The schedule of spectrum profiles available at a particular geolocation.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

PawGetSpectrumBatchCall
The Google Spectrum Database does not support batch requests, so this method always yields an UNIMPLEMENTED error.

PawGetSpectrumCall
Requests information about the available spectrum for a device at a location. Requests from a fixed-mode device must include owner information so the device can be registered with the database.

PawInitCall
Initializes the connection between a white space device and the database.

PawMethods
A builder providing access to all methods supported on paw resources. It is not used directly, but through the Spectrum hub.

PawNotifySpectrumUseCall
Notifies the database that the device has selected certain frequency ranges for transmission. Only to be invoked when required by the regulator. The Google Spectrum Database does not operate in domains that require notification, so this always yields an UNIMPLEMENTED error.

PawRegisterCall
The Google Spectrum Database implements registration in the getSpectrum method. As such this always returns an UNIMPLEMENTED error.

PawVerifyDeviceCall
Validates a device for white space use in accordance with regulatory rules. The Google Spectrum Database does not support master/slave configurations, so this always yields an UNIMPLEMENTED error.

PawsGetSpectrumBatchRequest
The request message for a batch available spectrum query protocol.

PawsGetSpectrumBatchResponse
The response message for the batch available spectrum query contains a schedule of available spectrum for the device at multiple locations.

PawsGetSpectrumRequest
The request message for the available spectrum query protocol which must include the device's geolocation.

PawsGetSpectrumResponse
The response message for the available spectrum query which contains a schedule of available spectrum for the device.

PawsInitRequest
The initialization request message allows the master device to initiate exchange of capabilities with the database.

PawsInitResponse
The initialization response message communicates database parameters to the requesting device.

PawsNotifySpectrumUseRequest
The spectrum-use notification message which must contain the geolocation of the Device and parameters required by the regulatory domain.

PawsNotifySpectrumUseResponse
An empty response to the notification.

PawsRegisterRequest
The registration request message contains the required registration parameters.

PawsRegisterResponse
The registration response message simply acknowledges receipt of the request and is otherwise empty.

PawsVerifyDeviceRequest
The device validation request message.

PawsVerifyDeviceResponse
The device validation response message.

RulesetInfo
This contains parameters for the ruleset of a regulatory domain that is communicated using the initialization and available-spectrum processes.

Spectrum
Central instance to access all Spectrum related resource activities

SpectrumMessage
Available spectrum can be logically characterized by a list of frequency ranges and permissible power levels for each range.

SpectrumSchedule
The spectrum schedule element combines an event time with spectrum profile to define a time period in which the profile is valid.

Vcard
A vCard-in-JSON message that contains only the fields needed for PAWS: - fn: Full name of an individual - org: Name of the organization - adr: Address fields - tel: Telephone numbers - email: Email addresses

VcardAddress
The structure used to represent a street address.

VcardTelephone
The structure used to represent a telephone number.

VcardTypedText
The structure used to represent an organization and an email address.

Enums

Error

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.